feature/remove-typescript-from-keystone
removing typescript from keystone

build issue with exporting ListConfig from Keystonejs types with typescript while trying to build for production
