--- backend/declaration.d.ts
@@ -1,5 +1,0 @@
-declare module '@keystone-6/auth';
-declare module '@keystone-6/core';
-declare module '@keystone-6/cloudinary';
-declare module '@keystone-6/fields-document';
-declare module 'graphql';
