+++ backend/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "backend",
-  "version": "0.4.0",
+  "name": "keystone-app",
+  "version": "1.0.2",
   "private": true,
   "scripts": {
     "dev": "keystone dev",
@@ -10,7 +10,6 @@
   },
   "dependencies": {
     "@keystone-6/auth": "^7.0.0",
-    "@keystone-6/cloudinary": "^7.0.3",
     "@keystone-6/core": "^5.0.0",
     "@keystone-6/fields-document": "^7.0.0",
     "typescript": "^4.9.5"
